Fix for PAC-3 lag
I have found the problem. It seems that the PAC-3 is using the only part using the largeExhaust model and it is clearly too GPU intensive. I have changed the exhaust to the medium one and it works perfect with no lag.
Evidence here;
Before: https://youtu.be/LT5TQ8t16eA
After: https://youtu.be/Zfh3cEQqL3o